task 4

+++ src/by/bsuir/webtech/lab1/task2/Task2.java
@@ -6,9 +6,9 @@ public class Task2 {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.print("x: ");
-        double x = input.nextInt();
+        int x = input.nextInt();
         System.out.print("y: ");
-        double y = input.nextInt();
+        int y = input.nextInt();
         System.out.print(0 <= y && y <= 5 && -4 <= x && x <= 4 || -3 <= y && y <= 0 && -6 <= x && x <= 6);
     }
 }
